#1352 refactor: move `model/` endpoints to own namespace / route prefix
