+++ plugins/BEdita/API/src/Controller/Model/PropertyTypesController.php
@@ -10,34 +10,30 @@
  *
  * See LICENSE.LGPL or <http://gnu.org/licenses/lgpl-3.0.html> for more details.
  */
-namespace BEdita\API\Controller;
+
+namespace BEdita\API\Controller\Model;
 
 /**
- * Controller for `/model/relations` endpoint.
+ * Controller for `/model/property_types` endpoint.
  *
  * @since 4.0.0
  *
- * @property \BEdita\Core\Model\Table\RelationsTable $Relations
+ * @property \BEdita\Core\Model\Table\PropertyTypesTable $PropertyTypes
  */
-class RelationsController extends ResourcesController
+class PropertyTypesController extends ModelController
 {
-    /**
-     * {@inheritDoc}
-     */
-    public $modelClass = 'Relations';
 
     /**
      * {@inheritDoc}
      */
-    protected $routeNamePrefix = 'api:model';
+    public $modelClass = 'PropertyTypes';
 
     /**
      * {@inheritDoc}
      */
     protected $_defaultConfig = [
         'allowedAssociations' => [
-            'left_object_types' => ['object_types'],
-            'right_object_types' => ['object_types'],
+            'properties' => ['properties'],
         ],
     ];
 }
