+++ plugins/BEdita/API/tests/TestCase/Controller/Model/PropertiesControllerTest.php
@@ -10,13 +10,14 @@
  *
  * See LICENSE.LGPL or <http://gnu.org/licenses/lgpl-3.0.html> for more details.
  */
-namespace BEdita\API\Test\TestCase\Controller;
+
+namespace BEdita\API\Test\TestCase\Controller\Model;
 
 use BEdita\API\TestSuite\IntegrationTestCase;
 use Cake\ORM\TableRegistry;
 
 /**
- * @coversDefaultClass \BEdita\API\Controller\PropertiesController
+ * @coversDefaultClass \BEdita\API\Controller\Model\PropertiesController
  */
 class PropertiesControllerTest extends IntegrationTestCase
 {
@@ -28,6 +29,8 @@ class PropertiesControllerTest extends IntegrationTestCase
     public $fixtures = [
         'plugin.BEdita/Core.property_types',
         'plugin.BEdita/Core.properties',
+        'plugin.BEdita/Core.locations',
+        'plugin.BEdita/Core.media',
     ];
 
     /**
